+++ xingyun-sc/src/main/java/com/lframework/xingyun/sc/entity/SaleOutSheetDetail.java
@@ -39,7 +39,7 @@ public class SaleOutSheetDetail extends BaseEntity implements BaseDto {
   /**
    * 出库数量
    */
-  private Integer orderNum;
+  private BigDecimal orderNum;
 
   /**
    * 原价
@@ -89,7 +89,7 @@ public class SaleOutSheetDetail extends BaseEntity implements BaseDto {
   /**
    * 已退货数量
    */
-  private Integer returnNum;
+  private BigDecimal returnNum;
 
   /**
    * 组合商品原始明细ID
