+++ xingyun-sc/src/main/java/com/lframework/xingyun/sc/entity/SaleReturnDetail.java
@@ -39,7 +39,7 @@ public class SaleReturnDetail extends BaseEntity implements BaseDto {
   /**
    * 退货数量
    */
-  private Integer returnNum;
+  private BigDecimal returnNum;
 
   /**
    * 原价
