+++ xingyun-sc/src/main/java/com/lframework/xingyun/sc/service/sale/SaleOutSheetDetailLotService.java
@@ -3,6 +3,7 @@ package com.lframework.xingyun.sc.service.sale;
 import com.lframework.starter.web.core.service.BaseMpService;
 import com.lframework.xingyun.sc.dto.sale.out.SaleOutSheetDetailLotDto;
 import com.lframework.xingyun.sc.entity.SaleOutSheetDetailLot;
+import java.math.BigDecimal;
 
 public interface SaleOutSheetDetailLotService extends BaseMpService<SaleOutSheetDetailLot> {
 
@@ -20,7 +21,7 @@ public interface SaleOutSheetDetailLotService extends BaseMpService<SaleOutSheet
    * @param id
    * @param num
    */
-  void addReturnNum(String id, Integer num);
+  void addReturnNum(String id, BigDecimal num);
 
   /**
    * 减少退货数量
@@ -28,5 +29,5 @@ public interface SaleOutSheetDetailLotService extends BaseMpService<SaleOutSheet
    * @param id
    * @param num
    */
-  void subReturnNum(String id, Integer num);
+  void subReturnNum(String id, BigDecimal num);
 }
